updating bootstrap

+++ bootstrap.sh
@@ -1,8 +1,7 @@
 #!/bin/bash
+KEEPSTAR_DOCKER_INSTALL=/opt/keepstar_docker
 
 set -e
-
-KEEPSTAR_DOCKER_INSTALL=/opt/keepstar_docker
 
 #check to see if running as root
 
@@ -36,22 +35,27 @@ echo "Ensuring $KEEPSTAR_DOCKER_INSTALL is available..."
 mkdir -p $KEEPSTAR_DOCKER_INSTALL
 cd $KEEPSTAR_DOCKER_INSTALL
 
-echo "Grabbing dockerfile and config file"
+    if [ ! -f /root/.keepstar-installed ]; then
 
-curl -L https://raw.githubusercontent.com/MurkyOwl/Keepstar-Docker/master/Dockerfile -o $KEEPSTAR_DOCKER_INSTALL/Dockerfile
-curl -L https://raw.githubusercontent.com/MurkyOwl/Keepstar-Docker/master/config.php -o $KEEPSTAR_DOCKER_INSTALL/config.php
-curl -L https://raw.githubusercontent.com/MurkyOwl/Keepstar-Docker/master/keepstar.conf -o $KEEPSTAR_DOCKER_INSTALL/keepstar.conf
+        echo "Grabbing dockerfile and config file"
 
-echo "Please edit your config files in /opt/keepstar_docker/ before continuing, Yes when ready."
-select yn in "Yes" "No"; do
-    case $yn in
-        Yes ) make install; break;;
-        No ) exit;;
-    esac
-done
+        curl -L https://raw.githubusercontent.com/MurkyOwl/Keepstar-Docker/master/Dockerfile -o $KEEPSTAR_DOCKER_INSTALL/Dockerfile
+        curl -L https://raw.githubusercontent.com/MurkyOwl/Keepstar-Docker/master/config.php -o $KEEPSTAR_DOCKER_INSTALL/config.php
+        curl -L https://raw.githubusercontent.com/MurkyOwl/Keepstar-Docker/master/keepstar.conf -o $KEEPSTAR_DOCKER_INSTALL/keepstar.conf
+
+        echo "Please edit your config files in /opt/keepstar_docker/ before continuing, Yes when ready."
+        select yn in "Yes" "No"; do
+            case $yn in
+                Yes ) make install; break;;
+                No ) exit;;
+            esac
+        done
+Mark it as installed
+    touch /root/.keepstar-installed
+    fi
 
 echo "Starting docker image.\n"
-docker run -d murkyowl:keepstar 
+docker build -t murkyowl:keepstar . && docker run -d -p 80:80 murkyowl:keepstar 
 
 
 echo "Docker has Started!"
